Removed an explicit typing which caused a compilation error.
I didn't catch it previously because it seems that that specific call was not possible dew to some JavaFX thing, but because the needed compilation was not redone, the error didn't appear.  It only appeared if I first did a `lein clean` before doing a `lein run`.

Note (to self and others): Always do a `lein clean`, then a `lein uberjar` before checking in any code.